requirements.txt
numpy~=1.25.2
pandas~=2.0.3
spacy~=3.5.3
streamlit~=1.27.0
nltk~=3.8.1
openai~=0.27.4
requests~=2.31.0
beautifulsoup4~=4.12.2
scikit-learn~=1.3.0
python-dotenv~=0.21.0
tenacity~=8.2.2